enhancement/API-SERVICES-Divide-api - Divide api on public and private - Add file interfaces - Add file constants - Refactor code
