+++ src/layouts/app/administrative/users/list/controller.ts
@@ -5,7 +5,7 @@ import { call, put, select, takeEvery } from 'redux-saga/effects';
 import { ActionCreator, ActionCreators, Controller, create } from 'redux-saga-controller';
 
 // services
-import { instanceAPI } from 'services/api.service';
+import { instanceAPI } from 'services/api-private.service';
 
 // constants
 import * as ROUTES from 'constants/routes';
